fix(demos/custom-ui): route context-menu decisions through shared store
